cmds/cpu: get it going on osx
